# packages/contracts/schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "TECHZONE Public & Admin API", "version": "1.0.0" },
  "servers": [{ "url": "/api" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "cookieAuth": { "type": "apiKey", "in": "cookie", "name": "tz_access" }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": ["code", "message", "requestId", "timestamp"],
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" },
          "requestId": { "type": "string" },
          "details": {},
          "timestamp": { "type": "string", "format": "date-time" }
        }
      },
      "User": {
        "type": "object",
        "required": ["id", "email", "name", "role"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "name": { "type": "string" },
          "role": { "type": "string" },
          "adminRole": { "type": ["string", "null"] },
          "permissions": { "type": "array", "items": { "type": "string" } }
        }
      },
      "Session": {
        "type": "object",
        "required": ["user", "accessToken", "csrfToken"],
        "properties": {
          "user": { "$ref": "#/components/schemas/User" },
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" },
          "csrfToken": { "type": "string" }
        }
      },
      "Page": {
        "type": "object",
        "required": ["items", "page", "pageSize", "total", "pageCount"],
        "properties": {
          "items": { "type": "array", "items": {} },
          "page": { "type": "integer", "minimum": 1 },
          "pageSize": { "type": "integer", "minimum": 1, "maximum": 100 },
          "total": { "type": "integer", "minimum": 0 },
          "pageCount": { "type": "integer", "minimum": 1 }
        }
      },
      "ProductVariant": {
        "type": "object",
        "required": ["id", "sku", "listPrice", "salePrice", "status", "availableQty", "inStock"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "sku": { "type": "string" },
          "modelNumber": { "type": ["string", "null"] },
          "optionValues": { "type": "object", "additionalProperties": { "type": "string" } },
          "listPrice": { "type": "integer", "minimum": 0 },
          "salePrice": { "type": "integer", "minimum": 0 },
          "status": { "type": "string", "enum": ["active", "inactive", "sold_out"] },
          "availableQty": {
            "description": "중앙 출고 창고의 실시간 가용 재고. Inventory 조회 장애 시 null",
            "type": ["integer", "null"],
            "minimum": 0
          },
          "inStock": { "type": ["boolean", "null"] }
        }
      }
    },
    "parameters": {
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": false, "schema": { "type": "string", "maxLength": 128 } },
      "CsrfToken": { "name": "X-CSRF-Token", "in": "header", "required": false, "schema": { "type": "string" } }
    },
    "responses": {
      "Error": { "description": "표준 오류 응답", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } }
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["email", "password"], "properties": { "email": { "type": "string" }, "password": { "type": "string" } } } } } },
        "responses": { "200": { "description": "로그인 성공", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Session" } } } }, "401": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/auth/refresh": {
      "post": { "operationId": "refreshSession", "responses": { "200": { "description": "토큰 회전", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Session" } } } }, "401": { "$ref": "#/components/responses/Error" } } }
    },
    "/products": {
      "get": {
        "operationId": "listProducts",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 48 } },
          { "name": "q", "in": "query", "schema": { "type": "string", "maxLength": 120 } },
          { "name": "search", "in": "query", "deprecated": true, "schema": { "type": "string", "maxLength": 120 } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "brand", "in": "query", "schema": { "type": "string" } },
          { "name": "minPrice", "in": "query", "schema": { "type": "integer", "minimum": 0 } },
          { "name": "maxPrice", "in": "query", "schema": { "type": "integer", "minimum": 0 } },
          { "name": "inStock", "in": "query", "schema": { "type": "boolean" } },
          { "name": "discounted", "in": "query", "schema": { "type": "boolean" } },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["popular", "newest", "discount", "price_asc", "price_desc"] } }
        ],
        "responses": { "200": { "description": "상품 목록", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Page" } } } } }
      }
    },
    "/products/by-slug/{slug}": {
      "get": {
        "operationId": "getProductBySlug",
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "variant별 가용 재고를 포함한 상품 상세",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["id", "slug", "name", "variants"],
                  "properties": {
                    "id": { "type": "string", "format": "uuid" },
                    "slug": { "type": "string" },
                    "name": { "type": "string" },
                    "variants": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ProductVariant" }
                    }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/checkout/quote": {
      "post": {
        "operationId": "createCheckoutQuote",
        "responses": {
          "200": { "description": "서버 가격과 variant 가용 재고 검증 완료" },
          "409": {
            "description": "품절 또는 요청 수량 초과(INSUFFICIENT_STOCK)",
            "$ref": "#/components/responses/Error"
          },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/wishlists/{userId}": {
      "get": {
        "operationId": "getWishlist",
        "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "본인 찜 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["items"],
                  "properties": { "items": { "type": "array", "items": {} } }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/wishlists/{userId}/{productId}": {
      "post": {
        "operationId": "addWishlistProduct",
        "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "productId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "$ref": "#/components/parameters/CsrfToken" }
        ],
        "responses": {
          "201": { "description": "찜 상품 추가" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "operationId": "removeWishlistProduct",
        "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "productId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "$ref": "#/components/parameters/CsrfToken" }
        ],
        "responses": {
          "204": { "description": "찜 상품 제거" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/orders": {
      "post": {
        "operationId": "createOrder",
        "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }, { "$ref": "#/components/parameters/CsrfToken" }],
        "responses": { "201": { "description": "주문 생성" }, "409": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/payments/confirm": {
      "post": {
        "operationId": "confirmPayment",
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["orderId", "amount"],
                "properties": {
                  "orderId": { "type": "string", "format": "uuid" },
                  "amount": { "type": "integer", "minimum": 0 },
                  "paymentKey": { "type": "string" },
                  "provider": { "type": "string", "enum": ["card", "kakaopay", "naverpay", "bank"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Order 원본 금액과 상품을 검증한 결제 승인" },
          "404": { "$ref": "#/components/responses/Error" },
          "409": {
            "description": "주문 합계 불일치(PAYMENT_AMOUNT_MISMATCH) 또는 결제 불가 상태(ORDER_NOT_PAYABLE)",
            "$ref": "#/components/responses/Error"
          },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/inventory/variants/{variantId}": {
      "patch": {
        "operationId": "adjustVariantInventory",
        "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
        "parameters": [
          { "name": "variantId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "$ref": "#/components/parameters/IdempotencyKey" },
          { "$ref": "#/components/parameters/CsrfToken" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["availableQty", "warehouseId", "reason"],
                "properties": {
                  "availableQty": { "type": "integer", "minimum": 0 },
                  "warehouseId": { "type": "string", "format": "uuid" },
                  "reason": { "type": "string", "maxLength": 500 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "variant·창고 단위 재고 조정" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/admin/dashboard": {
      "get": { "operationId": "getAdminDashboard", "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }], "responses": { "200": { "description": "운영 KPI" }, "403": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/dead-letters": {
      "get": { "operationId": "listDeadLetters", "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }], "responses": { "200": { "description": "DLQ 목록", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Page" } } } } } }
    },
    "/admin/system-status": {
      "get": {
        "operationId": "getSystemStatus",
        "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
        "responses": {
          "200": { "description": "전체 서비스 신뢰성 상태와 이벤트 처리 지표" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/admin/outbox": {
      "get": {
        "operationId": "listPendingOutbox",
        "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "minimum": 10, "maximum": 100 } },
          { "name": "q", "in": "query", "schema": { "type": "string" } },
          { "name": "service", "in": "query", "schema": { "type": "string" } },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["service", "event_type", "attempts", "occurred_at"] } },
          { "name": "direction", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } }
        ],
        "responses": {
          "200": { "description": "서비스별 미발행 Outbox 이벤트", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Page" } } } },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/admin/reservations": {
      "get": {
        "operationId": "listInventoryReservations",
        "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "minimum": 10, "maximum": 100 } },
          { "name": "q", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["reserved", "confirmed", "committed", "released"] } },
          { "name": "warehouseId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["created_at", "updated_at", "expires_at", "quantity", "status"] } },
          { "name": "direction", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } }
        ],
        "responses": {
          "200": { "description": "주문별 재고 예약 수명주기 원장", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Page" } } } },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    }
  }
}